+++ app/controllers/comments_controller.rb
@@ -7,7 +7,8 @@ class CommentsController < ApplicationController
 
   def create
     ParticipationFactory.create(task: comment.task, assignee: current_user, assigner: current_user)
-    respond_with comment if CommentLookManager.sync_comment(comment)
+    CommentLookManager.sync_comment(comment) unless current_user.site_admin?
+    respond_with comment
   end
 
   def show
@@ -25,13 +26,13 @@ class CommentsController < ApplicationController
       if params[:id].present?
         Comment.find(params[:id])
       else
-        task.comments.build(comment_params)
+        task.comments.build(comment_params.merge!(commenter_id: current_user.id))
       end
     end
   end
 
   def comment_params
-    params.require(:comment).permit(:commenter_id, :body, :task_id)
+    params.require(:comment).permit(:body, :task_id)
   end
 
   def render_404