+++ spec/features/journal_administration_spec.rb
@@ -108,11 +108,13 @@ feature "Journal Administration", js: true do
 
       describe do
         it "show Journal logo" do
-          with_aws_cassette(:yeti_image) do
-            journal.update_attributes(logo: File.open("spec/fixtures/yeti.jpg"))
-            visit "/admin/journals/1/roles/1/flow_manager"
-            find(".control-bar-link-icon").click
-            expect(page.find(".column-title-wrapper")).to have_css("img")
+          using_wait_time 5 do
+            with_aws_cassette(:yeti_image) do
+              journal.update_attributes(logo: File.open("spec/fixtures/yeti.jpg"))
+              visit "/admin/journals/1/roles/1/flow_manager"
+              find(".control-bar-link-icon").click
+              expect(page.find(".column-title-wrapper")).to have_css("img")
+            end
           end
         end
       end